Refactored Solution presenter
