+++ lab-2/src/jvmMain/kotlin/io/github/e1turin/output/view/features/present_solution/ui/method/SimpleIterationSystemSolution.kt
@@ -1,0 +1,2 @@
+package io.github.e1turin.output.view.features.present_solution.ui.method
+
